# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ESD Ticket Booking Platform API
  description: |
    Microservices-based event ticket booking platform for IS213.

    All responses follow a standard envelope:
    - **Success**: `{"code": N, "data": {...}}`
    - **Error**: `{"code": N, "message": "..."}`

    All endpoints accept an optional `X-Correlation-ID` header (UUID). If not provided, one is auto-generated for distributed tracing.
  version: 1.0.0
  contact:
    name: ESD Team
servers:
  - url: http://localhost:8000
    description: Kong API Gateway (local development)
...
paths:
...
  /api/notifications/user/{user_id}:
    get:
      tags: [Notifications]
      summary: Get notification history
      operationId: getUserNotifications
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID (Firebase UID)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of notifications to return
      responses:
        "200":
          description: Notification history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Notification"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/notifications/otp/send:
    post:
      tags: [Notifications]
      summary: Send OTP via SMS
      description: Sends a one-time password to the given mobile number via Twilio Verify.
      operationId: sendOtp
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SendOtpResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/notifications/otp/verify:
    post:
      tags: [Notifications]
      summary: Verify OTP code
      operationId: verifyOtp
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: OTP verification result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────────────
  # Tickets
  # ─────────────────────────────────────────────
...
components:
  parameters:
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Correlation ID for distributed tracing. Auto-generated if not provided.
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"

    # ── Event ───────────────────────────────────
...
    Notification:
      type: object
      properties:
        notification_id:
          type: integer
        user_id:
          type: string
        type:
          type: string
        channel:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    SendOtpRequest:
      type: object
      required: [mobile]
      properties:
        mobile:
          type: string
          description: Mobile number in E.164 format
          example: "+6591234567"
    SendOtpResponse:
      type: object
      properties:
        verification_sid:
          type: string
          description: Twilio Verify SID for subsequent verification
          example: "VExxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
    VerifyOtpRequest:
      type: object
      required: [verification_sid, code]
      properties:
        verification_sid:
          type: string
          example: "VExxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
        code:
          type: string
          description: 6-digit OTP code
          example: "123456"

    # ── Ticket ──────────────────────────────────
...